+++ src/components/SingleList.jsx
@@ -1,8 +1,11 @@
+import { useNavigate } from 'react-router-dom';
 import './SingleList.css';
 
 export function SingleList({ name, path, setListPath }) {
+	const navigate = useNavigate();
 	function handleClick() {
 		setListPath(path);
+		navigate('/list');
 	}
 
 	return (
